+++ src/components/player-header.ts
@@ -108,15 +108,12 @@ class PlayerHeader extends LitElement {
       css`
 
       .player-header-container {
-        margin-top: 0rem;
-        padding: 0.5rem;
-        padding-top: 0rem;
-        padding-bottom: 0rem;
-        max-width: 40rem;
+        max-width: var(--stpc-player-controls-max-width, 40rem);
         text-align: center;
         background-position: center;
         background-repeat: no-repeat;
         background-size: contain;
+        box-sizing: border-box;
         /*border: 1px solid red;  /*  FOR TESTING CONTROL LAYOUT CHANGES */
       }