+++ CHANGELOG.md
@@ -6,6 +6,15 @@ Change are listed in reverse chronological order (newest to oldest).
 
 <span class="changelog">
 
+###### [ 1.0.33 ] - 2026/02/18
+  
+  * Added theme variable `--stpc-player-controls-max-width` and `playerControlsMaxWidth` card configuration option; Max width value of player control areas (progress, transport deck, and volume details).
+  * Added theme variable `--stpc-player-controls-icon-justify-content` and `playerControlsIconJustifyContent` card configuration option; Justify content value of icons in the Player control areas (default is 'center').  Examples: left, center, right, space-around, space-between, space-evenly, etc.
+  * Added theme variable `--stpc-player-progress-label-font-size` and `playerProgressLabelFontSize` card configuration option; Font size of the player progressbar labels (default `x-small`).
+  * Added theme variable `--stpc-player-volume-label-font-size` and `playerVolumeLabelFontSize` card configuration option; Font size of the player volume labels (default `x-small`).
+  * Added theme variable `--stpc-player-progress-label-padding-lr` and `playerProgressLabelPaddingLR` card configuration option; Horizontal padding value of the player progress text labels (default is `0rem`).
+  * Added theme variable `--stpc-card-content-margin` and `cardContentSectionMargin` card configuration option; Card content section margin value (default is `0.0rem`).
+
 ###### [ 1.0.32 ] - 2026/01/08
   
   * Updated Recently Played section processing for devices that do not support websockets.